+++ src/index.ts
@@ -4,14 +4,6 @@ import unformatMoney from './utils/unformatMoney';
 import textEllipsis from './utils/textEllipsis';
 import formatCep from './utils/formatCep';
 
-import cnpjIsValid from './utils/cnpjIsValid';
-import cpfIsValid from './utils/cpfIsValid';
+import isCnpj from './utils/isCnpj';
 
-export {
-  formatMoney,
-  unformatMoney,
-  textEllipsis,
-  formatCep,
-  cnpjIsValid,
-  cpfIsValid,
-};
+export { formatMoney, unformatMoney, textEllipsis, formatCep, isCnpj, isCnpj };