+++ src/utils/cnpjIsValid/index.ts
@@ -1,3 +1,0 @@
-import cnpjIsValid from './cnpjIsValid';
-
-export default cnpjIsValid;