+++ src/utils/cpfIsValid/index.ts
@@ -1,3 +1,0 @@
-import cpfIsValid from './cpfIsValid';
-
-export default cpfIsValid;