+++ src/utils/isCnpj/index.ts
@@ -1,0 +1,3 @@
+import isCnpj from './isCnpj';
+
+export default isCnpj;