+++ src/utils/isCpf/index.ts
@@ -1,0 +1,3 @@
+import isCpf from './isCpf';
+
+export default isCpf;